Split pipeline-wait observability tests (#73)
Move the record-ordering contract out of the async-boundary suite so
each pipeline-wait test module stays below the repository's size limit.

Document the complete split suite and regenerate the wheel-manifest
snapshot for the new test module.
