+++ docs/developers-guide.md
@@ -541,16 +541,17 @@ covers the same ground until that surface exists.
 Three verification layers cover this seam; keep all three when changing it:
 
 ```bash
-# Hypothesis state machine over randomized completion orders, the pinned
-# boundary cases, the async boundary, the log records, and the correlation and
-# teardown-outcome fields. These are five modules by test concern, sharing
-# scaffolding via `cuprum/unittests/_pipeline_wait_support.py`.
+# The state machine, boundary cases, async boundary, observability, fail-fast
+# event, and end-to-end wiring are split into seven focused modules sharing
+# `cuprum/unittests/_pipeline_wait_support.py` scaffolding.
 uv run pytest -q \
+  cuprum/unittests/test_pipeline_wait.py \
   cuprum/unittests/test_pipeline_wait_state_machine.py \
   cuprum/unittests/test_pipeline_wait_examples.py \
   cuprum/unittests/test_pipeline_wait_async.py \
   cuprum/unittests/test_pipeline_wait_observability.py \
-  cuprum/unittests/test_pipeline_wait_correlation.py
+  cuprum/unittests/test_pipeline_wait_fail_fast_event.py \
+  cuprum/unittests/test_pipeline_fail_fast_wiring.py
 # CrossHair PEP 316 contracts over the bounded symbolic model.
 uv run pytest -q cuprum/unittests/test_pipeline_wait_crosshair.py -m crosshair
 uv run crosshair check \
